# packages/studio/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InkOS Studio API",
    "version": "1.5.0",
    "description": "Auto-generated OpenAPI spec from 157 endpoints across 33 route modules."
  },
  "servers": [
    {
      "url": "http://localhost:4579",
      "description": "InkOS Studio API server"
    }
  ],
  "paths": {
...
    "/api/v1/style/analyze": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/analyze",
        "operationId": "post__api_v1_style_analyze",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/diagnostics": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/diagnostics",
        "operationId": "post__api_v1_style_diagnostics",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/ai-tells": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/ai-tells",
        "operationId": "post__api_v1_style_ai_tells",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/compare": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/compare",
        "operationId": "post__api_v1_style_compare",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/adjustments/plan": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/adjustments/plan",
        "operationId": "post__api_v1_style_adjustments_plan",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/adjustments/preview": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/adjustments/preview",
        "operationId": "post__api_v1_style_adjustments_preview",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/style/import": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/books/:id/style/import",
        "operationId": "post__api_v1_books_id_style_import",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/extract-text": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/extract-text",
        "operationId": "post__api_v1_style_extract_text",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/import-url": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/import-url",
        "operationId": "post__api_v1_style_import_url",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/preprocess": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/preprocess",
        "operationId": "post__api_v1_style_preprocess",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/relayout": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/relayout",
        "operationId": "post__api_v1_style_relayout",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/preprocess/inspect": {
      "post": {
        "tags": [
          "style"
        ],
        "summary": "POST /api/v1/style/preprocess/inspect",
        "operationId": "post__api_v1_style_preprocess_inspect",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}